only fill edittext
